+++ http-client/Network/HTTP/Client/Body.hs
@@ -14,7 +14,7 @@ module Network.HTTP.Client.Body
 
 import Network.HTTP.Client.Connection
 import Network.HTTP.Client.Types
-import Control.Exception (throwIO, assert)
+import Control.Exception (assert)
 import Data.ByteString (ByteString, empty, uncons)
 import Data.IORef
 import qualified Data.ByteString as S
@@ -88,7 +88,7 @@ makeGzipReader brRead = do
                         else do
                             writeIORef istate Nothing
                             return bs
-                Z.PRError e -> throwIO $ HttpZlibException e
+                Z.PRError e -> throwHttp $ HttpZlibException e
         start = do
             bs <- brRead
             if S.null bs
@@ -119,7 +119,7 @@ makeLengthReader count0 Connection {..} = do
             then return empty
             else do
                 bs <- connectionRead
-                when (S.null bs) $ throwIO $ ResponseBodyTooShort (fromIntegral count0) (fromIntegral $ count0 - count)
+                when (S.null bs) $ throwHttp $ ResponseBodyTooShort (fromIntegral count0) (fromIntegral $ count0 - count)
                 case compare count $ S.length bs of
                     LT -> do
                         let (x, y) = S.splitAt count bs
@@ -162,7 +162,7 @@ makeChunkedReader raw conn@Connection {..} = do
     readChunk 0 = return (empty, 0)
     readChunk remainder = do
         bs <- connectionRead
-        when (S.null bs) $ throwIO InvalidChunkHeaders
+        when (S.null bs) $ throwHttp InvalidChunkHeaders
         case compare remainder $ S.length bs of
             LT -> do
                 let (x, y) = S.splitAt remainder bs
@@ -180,12 +180,12 @@ makeChunkedReader raw conn@Connection {..} = do
 
     requireNewline = do
         bs <- connectionReadLine conn
-        unless (S.null bs) $ throwIO InvalidChunkHeaders
+        unless (S.null bs) $ throwHttp InvalidChunkHeaders
 
     readHeader = do
         bs <- connectionReadLine conn
         case parseHex bs of
-            Nothing -> throwIO InvalidChunkHeaders
+            Nothing -> throwHttp InvalidChunkHeaders
             Just hex -> return (bs `S.append` "\r\n", hex)
 
     parseHex bs0 =