+++ http-client/Network/HTTP/Client/Headers.hs
@@ -6,7 +6,6 @@ module Network.HTTP.Client.Headers
     ) where
 
 import           Control.Applicative            ((<$>), (<*>))
-import           Control.Exception              (throwIO)
 import           Control.Monad
 import qualified Data.ByteString                as S
 import qualified Data.ByteString.Char8          as S8
@@ -32,7 +31,7 @@ parseStatusHeaders conn timeout' cont
   where
     withTimeout = case timeout' of
         Nothing -> id
-        Just  t -> timeout t >=> maybe (throwIO ResponseTimeout) return
+        Just  t -> timeout t >=> maybe (throwHttp ResponseTimeout) return
 
     getStatus = withTimeout next
       where
@@ -55,7 +54,7 @@ parseStatusHeaders conn timeout' cont
         -- Ensure that there is some data coming in. If not, we want to signal
         -- this as a connection problem and not a protocol problem.
         bs <- connectionRead conn
-        when (S.null bs) $ throwIO NoResponseDataReceived
+        when (S.null bs) $ throwHttp NoResponseDataReceived
         connectionReadLineWith conn bs >>= parseStatus 3
 
     parseStatus :: Int -> S.ByteString -> IO (Status, HttpVersion)
@@ -66,7 +65,7 @@ parseStatusHeaders conn timeout' cont
             msg = S.dropWhile (== charSpace) bs3
         case (,) <$> parseVersion ver <*> readInt code of
             Just (ver', code') -> return (Status code' msg, ver')
-            Nothing -> throwIO $ InvalidStatusLine bs
+            Nothing -> throwHttp $ InvalidStatusLine bs
 
     stripPrefixBS x y
         | x `S.isPrefixOf` y = Just $ S.drop (S.length x) y
@@ -81,7 +80,7 @@ parseStatusHeaders conn timeout' cont
             Just (i, "") -> Just i
             _ -> Nothing
 
-    parseHeaders 100 _ = throwIO OverlongHeaders
+    parseHeaders 100 _ = throwHttp OverlongHeaders
     parseHeaders count front = do
         line <- connectionReadLine conn
         if S.null line
@@ -93,7 +92,7 @@ parseStatusHeaders conn timeout' cont
     parseHeader :: S.ByteString -> IO Header
     parseHeader bs = do
         let (key, bs2) = S.break (== charColon) bs
-        when (S.null bs2) $ throwIO $ InvalidHeader bs
+        when (S.null bs2) $ throwHttp $ InvalidHeader bs
         return (CI.mk $! strip key, strip $! S.drop 1 bs2)
 
     strip = S.dropWhile (== charSpace) . fst . S.spanEnd (== charSpace)